refactor(reactstrap-validation-select): move patient dropdown components
