+++ packages/reactstrap-validation-select/resources.js
@@ -1,5 +1,3 @@
-import React, { Fragment } from 'react';
-import { components } from 'react-select';
 import {
   avProvidersApi,
   avOrganizationsApi,
@@ -11,6 +9,11 @@ import {
   avThanosApi,
 } from '@availity/api-axios';
 import AvResourceSelect from './AvResourceSelect';
+import {
+  PatientOption,
+  PatientSingleValue,
+  PatientMultiValueLabel,
+} from './resource-components';
 
 const AvProviderSelect = AvResourceSelect.create({
   resource: avProvidersApi,
@@ -49,109 +52,6 @@ const AvCodeSelect = AvResourceSelect.create({
   requiredParams: ['list'],
   watchParams: ['list'],
 });
-
-/* eslint-disable react/prop-types */
-const PatientOption = props => {
-  const {
-    lastName,
-    firstName,
-    middleName,
-    payerName,
-    subscriberMemberId,
-    memberId,
-    birthDate: bd,
-  } = props.data;
-
-  const option = (
-    <Fragment>
-      <strong>
-        {lastName}, {firstName} {middleName ? `${middleName}.` : ''}
-      </strong>
-      &nbsp;&nbsp;&nbsp;{payerName} {memberId || subscriberMemberId}
-    </Fragment>
-  );
-
-  let birthDate = null;
-  // Format date
-  if (bd) {
-    birthDate = (
-      <Fragment>
-        &nbsp;&nbsp;&nbsp;&nbsp;{bd[5]}
-        {bd[6]}/{bd[8]}
-        {bd[9]}/{bd[0]}
-        {bd[1]}
-        {bd[2]}
-        {bd[3]}
-      </Fragment>
-    );
-  }
-
-  return (
-    <components.Option {...props}>
-      {option}
-      {birthDate}
-    </components.Option>
-  );
-};
-
-const PatientSingleValue = props => {
-  const {
-    lastName,
-    firstName,
-    middleName,
-    payerName,
-    subscriberMemberId,
-    memberId,
-    birthDate: bd,
-  } = props.data;
-
-  const value = (
-    <Fragment>
-      <strong>
-        {lastName}, {firstName} {middleName ? `${middleName}.` : ''}
-      </strong>
-      &nbsp;&nbsp;&nbsp;{payerName} {memberId || subscriberMemberId}
-    </Fragment>
-  );
-
-  let birthDate = null;
-  // Format date
-  if (bd) {
-    birthDate = (
-      <Fragment>
-        &nbsp;&nbsp;&nbsp;&nbsp;{bd[5]}
-        {bd[6]}/{bd[8]}
-        {bd[9]}/{bd[0]}
-        {bd[1]}
-        {bd[2]}
-        {bd[3]}
-      </Fragment>
-    );
-  }
-
-  return (
-    <components.SingleValue {...props}>
-      {value}
-      {birthDate}
-    </components.SingleValue>
-  );
-};
-
-const PatientMultiValueLabel = props => {
-  const { lastName, firstName, middleName } = props.data;
-
-  const value = (
-    <Fragment>
-      {lastName}, {firstName} {middleName ? `${middleName}.` : ''}
-    </Fragment>
-  );
-
-  return (
-    <components.MultiValueLabel {...props}>{value}</components.MultiValueLabel>
-  );
-};
-
-/* eslint-enable react/prop-types */
 
 const AvPatientSelect = AvResourceSelect.create({
   components: {
